Persist build dir across jobs
Fixes an issue where the deploy fails because `_site` doesn't exist.

This allows the `_site` directory to be built in the `build` job and then
restored for the `deploy` job.

The `persist_to_workspace` syntax is a little strange. The examples I've seen
suggest using a `working_directory` because otherwise you get some strange
behavior.