Dynamically generate a request form from agency metadata file (#102)
* Add stub metadata API

* Utility to convert agency metadata to JSON Schema

* Add stubbed dropdown with agencies

* Add flux stores and actions

* Move React components to dir

* Fetch agency, generate a form

* Move actions to directory

* Translate metadata to uiSchema

* Update dynamic form title

* Styling for FOIA request form

* Update canary

* Add React/Flux concepts to README

* Combine required and additional fields

It's simpler to keep these as a single list with a `required` property to
distinguish between them. This also allows us to specify the field ordering as
a single list.

* Handling for custom metadata types

* Use the agency name for the component name

This is a temporary hack to get OGE's dynamic form working. For centralized
agencies we just assume that the agency name matches the component name.