Continuous deployment for Acquia Cloud (#176)
* Split settings into more environments

`local` is for your local development environment. URLs should be to localhost
or to host file overrides that point to local VMs. This is the default.

'development`, `staging`, and `production` are for the Acquia Cloud
environments.

We currently don't have `production` setup, so point to staging for now.

* Add deploy script for Acquia Cloud

* Configure CircleCI to deploy to Acquia Cloud

* Remove Basic Auth

This is unsupported on the Acquia Cloud[1].

[1]: https://docs.acquia.com/article/password-protecting-nonproduction-environments#authuserfile

* Deploy master to development environment

* Deploy release- branches to staging environment

* Note on how branches are deployed

* Remove cloud.gov deploy configuration and files

* Revert "Remove cloud.gov deploy configuration and files"

This reverts commit 81e2b0ceaee75c4bda49016595681b065a107d8e.

* Add build environment for cloud.gov

This allows us to continue deploying to cloud.gov should we need any extra dev
environments.