Make supporting_docs the additional fields section
Instead of having a separate "additional fields" section, we want the supporting
docs section to contain the attachments field as well as the agency component
specific fields, aka the "10% fields".

We add a flag onto exactly _one_ of the existing sections. That section will be
made the "additional fields" section. Any field that doesn't match the
`fieldNames` of the existing sections should be added to the form under this
section.