FOIA-280: Handle agency value or error on change.
Adds a listener for the typeahead:change event that will find and set
the selected agency when the user focus leaves the typeahead field.
This will allow properly setting an agency when the input changes or
setting an error if the input has changed to a value with no valid
agency from the search.